refactor/#277 serviceType serviceProtocol로 네이밍 변경
