+++ KkuMulKum/Source/AddMeeting/InviteCode/ServiceProtocol/InviteCodeServiceProtocol.swift
@@ -1,0 +1,14 @@
+//
+//  InviteCodeServiceProtocol.swift
+//  KkuMulKum
+//
+//  Created by YOUJIM on 7/13/24.
+//
+
+import Foundation
+
+protocol InviteCodeServiceProtocol {
+    func joinMeeting(
+        with request: RegisterMeetingsModel
+    ) async throws -> ResponseBodyDTO<RegisterMeetingsResponseModel>?
+}
